+++ src/marshmallow/__init__.py
@@ -15,7 +15,7 @@ from marshmallow.decorators import (
 )
 from marshmallow.exceptions import ValidationError
 from marshmallow.schema import Schema, SchemaOpts
-from marshmallow.utils import EXCLUDE, INCLUDE, RAISE, missing, pprint
+from marshmallow.utils import EXCLUDE, INCLUDE, RAISE, missing
 
 from . import fields