+++ tests/test_schema.py
@@ -640,21 +640,6 @@ def test_can_serialize_time(user, serialized_user):
     assert serialized_user["time_registered"] == expected
 
 
-def test_json_module_is_deprecated():
-    with pytest.deprecated_call():
-
-        class UserJSONSchema(Schema):
-            name = fields.String()
-
-            class Meta:
-                json_module = mockjson
-
-    user = User("Joe")
-    s = UserJSONSchema()
-    result = s.dumps(user)
-    assert result == mockjson.dumps("val")
-
-
 def test_render_module():
     class UserJSONSchema(Schema):
         name = fields.String()
@@ -2244,21 +2229,6 @@ class TestSelfReference:
         assert data["employer"]["name"] == employer.name
         assert data["employer"]["age"] == employer.age
 
-    def test_nesting_schema_self_string(self, user, employer):
-        with pytest.warns(
-            DeprecationWarning, match="Passing 'self' to `Nested` is deprecated"
-        ):
-
-            class SelfSchema(Schema):
-                name = fields.String()
-                age = fields.Integer()
-                employer = fields.Nested("self", exclude=("employer",))
-
-        data = SelfSchema().dump(user)
-        assert data["name"] == user.name
-        assert data["employer"]["name"] == employer.name
-        assert data["employer"]["age"] == employer.age
-
     def test_nesting_within_itself_meta(self, user, employer):
         class SelfSchema(Schema):
             employer = fields.Nested(lambda: SelfSchema(exclude=("employer",)))
@@ -2303,25 +2273,6 @@ class TestSelfReference:
         relative = data["rels"][0]
         assert relative == user.relatives[0].name
 
-    def test_multiple_pluck_self_string(self, user):
-        with pytest.warns(
-            DeprecationWarning, match="Passing 'self' to `Nested` is deprecated"
-        ):
-
-            class MultipleSelfSchema(Schema):
-                emp = fields.Pluck("self", "name", attribute="employer")
-                rels = fields.Pluck("self", "name", many=True, attribute="relatives")
-
-                class Meta:
-                    fields = ("name", "emp", "rels")
-
-        schema = MultipleSelfSchema()
-        user.relatives = [User(name="Bar", age=12), User(name="Baz", age=34)]
-        data = schema.dump(user)
-        assert len(data["rels"]) == len(user.relatives)
-        relative = data["rels"][0]
-        assert relative == user.relatives[0].name
-
     def test_nested_self_many_lambda(self):
         class SelfManySchema(Schema):
             relatives = fields.Nested(lambda: SelfManySchema(), many=True)
@@ -2337,50 +2288,12 @@ class TestSelfReference:
         assert data["relatives"][0]["name"] == person.relatives[0].name
         assert data["relatives"][0]["age"] == person.relatives[0].age
 
-    def test_nested_self_many_string(self):
-        with pytest.warns(
-            DeprecationWarning, match="Passing 'self' to `Nested` is deprecated"
-        ):
-
-            class SelfManySchema(Schema):
-                relatives = fields.Nested("self", many=True)
-
-                class Meta:
-                    additional = ("name", "age")
-
-        person = User(name="Foo")
-        person.relatives = [User(name="Bar", age=12), User(name="Baz", age=34)]
-        data = SelfManySchema().dump(person)
-        assert data["name"] == person.name
-        assert len(data["relatives"]) == len(person.relatives)
-        assert data["relatives"][0]["name"] == person.relatives[0].name
-        assert data["relatives"][0]["age"] == person.relatives[0].age
-
     def test_nested_self_list(self):
         class SelfListSchema(Schema):
             relatives = fields.List(fields.Nested(lambda: SelfListSchema()))
 
             class Meta:
                 additional = ("name", "age")
-
-        person = User(name="Foo")
-        person.relatives = [User(name="Bar", age=12), User(name="Baz", age=34)]
-        data = SelfListSchema().dump(person)
-        assert data["name"] == person.name
-        assert len(data["relatives"]) == len(person.relatives)
-        assert data["relatives"][0]["name"] == person.relatives[0].name
-        assert data["relatives"][0]["age"] == person.relatives[0].age
-
-    def test_nested_self_list_string(self):
-        with pytest.warns(
-            DeprecationWarning, match="Passing 'self' to `Nested` is deprecated"
-        ):
-
-            class SelfListSchema(Schema):
-                relatives = fields.List(fields.Nested("self"))
-
-                class Meta:
-                    additional = ("name", "age")
 
         person = User(name="Foo")
         person.relatives = [User(name="Bar", age=12), User(name="Baz", age=34)]